Share above the bar; portrait Share goes straight to native sheet
Move the Share button above the triangle-strip bar (into the wordmark
row) on both the Knowledge Portrait card and the Recently Expanding tile.

Make the Knowledge Portrait Share button skip the preview modal and go
directly to the native share sheet. The portrait image is pre-captured in
the background (off-screen card) so the share fires synchronously inside
the tap gesture — iOS aborts a share if html2canvas is awaited in-handler.
If the user taps before capture finishes, it falls back to the existing
preview modal. Capture logic is extracted into useSharePortraitCapture and
reused by the modal.

Also render the triangle-strip graphic in the shared portrait image so the
exported PNG matches the on-screen card.
